+++ src/activitypub/identifiers.rs
@@ -5,6 +5,7 @@ use mitra_models::{
     posts::types::Post,
     profiles::types::DbActorProfile,
 };
+use mitra_utils::urls::get_hostname;
 
 use crate::errors::ValidationError;
 
@@ -79,6 +80,12 @@ pub fn local_emoji_id(instance_url: &str, emoji_name: &str) -> String {
 
 pub fn local_tag_collection(instance_url: &str, tag_name: &str) -> String {
     format!("{}/collections/tags/{}", instance_url, tag_name)
+}
+
+pub fn validate_object_id(object_id: &str) -> Result<(), ValidationError> {
+    get_hostname(object_id)
+        .map_err(|_| ValidationError("invalid object ID"))?;
+    Ok(())
 }
 
 pub fn parse_local_actor_id(